some fight

+++ src/include/skill_type.h
@@ -13,7 +13,7 @@ struct skill_t {
     mod_t cost;
     bullet sample;
     bool is_range;
-    int dmg, u_l, u_r, d_l, d_r;
+    float dmg, u_l, u_r, d_l, d_r;
     void in_damage(std::istream& stream);
 };
